# public/swagger-development.json
{
  "info": {
    "title": "API REST documentation",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "description": "API REST to manage users"
  },
  "basePath": "/",
  "host": "localhost:8080",
  "tags": [
    {
      "name": "members",
      "description": "Methods for managing users"
    },
    {
      "name": "members-id",
      "description": "Methods for managing users with ID"
    }
  ],
  "paths": {
    "/members/{id}": {
      "get": {
        "tags": ["members-id"],
        "summary": "members/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": ["members-id"],
        "summary": "members/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 36
              }
            },
            "required": true
          },
          {
            "name": "name",
            "in": "body",
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 255
              }
            },
            "required": true
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": ["members-id"],
        "summary": "members/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/members": {
      "get": {
        "tags": ["members"],
        "summary": "members",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "max",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": ["members"],
        "summary": "members",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 255
              }
            },
            "required": true
          }
        ],
        "responses": {}
      }
    }
  },
  "swagger": "2.0"
}